requirements.txt
customtkinter
Pillow
python-dateutil
plyer
bcrypt
tkcalendar
matplotlib